:recycle: Refactor transfers handlers structure
